perf(clustering) improve hash calculation performance
### Summary

With large configurations the hash calculation can take a lot of time
and memory. This PR optimizes the performance of configuration hash
calculation. The time is cut to about half of what it used to be and
the peak memory usage has also gone down.

This does not produce exactly same hash on certain string as it did
before (as I optimized array case).